+++ ip2geotools_locator/calculations/__init__.py
@@ -1,5 +1,4 @@
 """Module for importing calculation methods"""
 from ip2geotools_locator.calculations.average import Average
 from ip2geotools_locator.calculations.clustering import Clustering
-from ip2geotools_locator.calculations.confidence_interval import ConfidenceInterval
 from ip2geotools_locator.calculations.median import Median