+++ ip2geotools_locator/database_connectors/ip_city.py
@@ -28,39 +28,37 @@ class IpCityDB:
         try:
             # Try to get and return location
             self.__db_data = DbIpCity.get(ip_address)
-            logger.info("%s: DB returned location %.3f N, %.3f E",
-                        __name__, self.__db_data.latitude, self.__db_data.longitude)
+            if self.__db_data.latitude is None or self.__db_data.longitude is None:
+                raise InvalidResponseError
+            logger.info("DB returned location %.3f N, %.3f E", self.__db_data.latitude,
+                        self.__db_data.longitude)
             return Location(self.__db_data.latitude, self.__db_data.longitude)
 
         except IpAddressNotFoundError as exception:
             # Handling for IpAddressNotFoundError exception
-            logger.warning("%s: Database could not find IP address. IpAddressNotFoundError: %s ",
-                           __name__, str(exception))
+            logger.warning("Database could not find IP address. IpAddressNotFoundError: %s ",
+                           str(exception))
 
         except PermissionRequiredError as exception:
             # Handling for PermissionRequiredError exception
-            logger.critical("%s: Additional setings required for DB. PermissionRequiredError: %s ",
-                            __name__, str(exception))
+            logger.critical("Additional setings required for DB. PermissionRequiredError: %s ",
+                            str(exception))
 
         except ServiceError as exception:
             # Handling for ServiceError exception
-            logger.error(
-                "%s: Service is unavailable. ServiceError: %s ", __name__, str(exception))
+            logger.error("Service is unavailable. ServiceError: %s ", str(exception))
 
         except LimitExceededError as exception:
             # Handling for LimitExceededError exception
-            logger.warning("%s: LimitExceededError: %s ",
-                           __name__, str(exception))
+            logger.warning("LimitExceededError: %s ", str(exception))
 
-        except (LocationError, InvalidRequestError, InvalidResponseError) as exception:
+        except (LocationError, InvalidRequestError, InvalidResponseError, KeyError) as exception:
             # Handling for invalid data, request and response exception
-            logger.error("%s: returned %s ", __name__,
-                         str(exception.__class__))
+            logger.error("returned %s ", str(exception.__class__))
 
         except TypeError as exception:
             # Handling for TypeError exception (in case of database returning None values)
-            logger.warning(
-                "%s: DB returned invalid values. TypeError: %s ", __name__, str(exception))
+            logger.warning("DB returned invalid values. TypeError: %s ", str(exception))
 
     def add_to_map(self):
         """
@@ -68,15 +66,14 @@ class IpCityDB:
         Call get_location(ip) method before adding any markers to map
         """
         try:
-            logger.debug("%s: Calling add_marker method for %s DB",
-                         __name__, DbIpCity.__name__)
+            logger.debug("Calling add_marker method for %s DB", DbIpCity.__name__)
             self.m.add_marker_noncommercial(DbIpCity.__name__,
                                             self.__db_data.ip_address,
                                             self.__db_data.country,
+                                            self.__db_data.region,
                                             self.__db_data.city,
                                             self.__db_data.latitude,
                                             self.__db_data.longitude)
         except AttributeError as exception:
             # Handling for AttributeError exception (in case of database returning None values)
-            logger.warning("%s: Cannot add empty marker %s ",
-                           __name__, str(exception))
+            logger.warning("Cannot add empty marker %s ", str(exception))